server/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pyserial==3.5
pydantic==2.5.0
